Refactoring of code into common jar
