+++ build.gradle
@@ -9,7 +9,7 @@ buildscript {
     }
 }
 
-version "3.0.1-SNAPSHOT"
+version "3.0.2-SNAPSHOT"
 group "org.grails.plugins"
 
 apply plugin:"eclipse"
@@ -40,9 +40,7 @@ dependencies {
     console "org.grails:grails-console"
 
     /** dependencies particular to this plugin **/
-    compile "com.auth0:java-jwt:3.18.2"
-    compile "com.auth0:jwks-rsa:0.20.0"
-    compile "org.springframework.boot:spring-boot-starter-oauth2-client"
+    compile "au.org.ala:ala-ws-security:1.0-SNAPSHOT"
     /** end of dependencies particular to this plugin **/
 
     profile "org.grails.profiles:web-plugin"
@@ -51,11 +49,8 @@ dependencies {
     testCompile "org.grails.plugins:geb"
     testRuntime "org.seleniumhq.selenium:selenium-htmlunit-driver:2.47.1"
     testRuntime "net.sourceforge.htmlunit:htmlunit:2.18"
-//    testCompile "org.grails:grails-test-mixins:3.3.0"
     testCompile "org.grails:grails-gorm-testing-support:2.1.3"
     testCompile "org.grails:grails-web-testing-support:2.1.3"
-//    testCompile "org.grails:grails-logging"
-//        testCompile 'org.grails:grails-validation:3.3.14'
 }
 
 bootRun {
